move utils file to core dir
